requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Validación de datos
pydantic==2.5.0
pydantic-settings==2.1.0

# Computación científica
numpy==1.24.3
scipy==1.11.4

# Base de datos PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Utilidades web
python-multipart==0.0.6
python-dotenv==1.0.0

# Autenticación (opcional para futuro)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Desarrollo y testing
pytest==7.4.3
pytest-asyncio==0.21.1